project/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.2,<5.3
psycopg2>=2.9,<3.0
django-cors-headers>=3.14,<3.15
django-filter>=23.2,<24.0
Pillow>=10.2.0
python-decouple>=3.8
django-ses>=3.5.2